Separate single and paired end mapping metadata.
